# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tapio Observability Platform API
  description: |
    Tapio provides enterprise-grade observability with semantic correlation.
    
    ## Features
    - Real-time event streaming with OTEL context
    - AI-powered semantic correlation
    - Multi-dimensional root cause analysis
    - Predictive analytics and business impact assessment
    
    ## Authentication
    API keys or JWT tokens in Authorization header
    
    ## Rate Limiting
    - 10,000 requests per minute per API key
    - 165,000 events per second for streaming
  version: 1.0.0
  contact:
    name: Tapio Support
    email: support@tapio.io
    url: https://tapio.io/support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.tapio.io/api/v1
    description: Production server
  - url: https://staging-api.tapio.io/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /events/stream:
    get:
      tags:
        - Events
      summary: Stream real-time events (SSE)
      operationId: streamEvents
      parameters:
        - name: filter
          in: query
          description: JSON-encoded filter criteria
          schema:
            type: string
        - name: lookback
          in: query
          description: Duration to include historical events
          schema:
            type: string
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
  /correlations/analyze:
    post:
      tags:
        - Correlations
      summary: Analyze events for correlations
      operationId: analyzeCorrelations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/Event'
                analysis_type:
                  type: string
                  enum: [SEMANTIC, TEMPORAL, CAUSAL]
                options:
                  type: object
                  properties:
                    include_root_cause:
                      type: boolean
                    include_predictions:
                      type: boolean
                    confidence_threshold:
                      type: number
                      minimum: 0
                      maximum: 1
      responses:
        '202':
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        context:
          type: object
          properties:
            service:
              type: string
            namespace:
              type: string
            trace_id:
              type: string
            span_id:
              type: string
...
    CorrelationFinding:
      type: object
      properties:
        id:
          type: string
        pattern_type:
          type: string
        confidence:
          type: number
        timestamp:
          type: string
          format: date-time
        description:
          type: string
        related_event_ids:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
...
    AnalysisResponse:
      type: object
      properties:
        analysis_id:
          type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/CorrelationFinding'
        status:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        event_count:
          type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT